fix: rename builder v2 file (EC-166)
